Rename MainPage to App and move to shared
