+++ src/components/App/Importers/Importer.vue
@@ -53,7 +53,7 @@ import { mapActions } from 'vuex';
 import { remote } from 'electron';
 import { REMOVE_IMPORTER_ACTION } from '@/store/modules/Config';
 import { scrape } from '@/modules/scrapers';
-import DeleteImporterDialog from '@/components/MainPage/Importers/DeleteImporterDialog';
+import DeleteImporterDialog from './DeleteImporterDialog';
 
 export default {
   components: {
