+++ src/components/shared/ProfileChip.vue
